+++ include/kaguya/math/HemiUniformPdf.h
@@ -14,7 +14,7 @@ namespace kaguya {
         public:
             double pdf(const Vector3 &inDir, const Vector3 &normal, const Vector3 &outDir) override;
 
-            Vector3 random(const Vector3 &inDir, const Vector3 &normal, float &samplePdf) override;
+            Vector3 random(const Vector3 &inDir, const Vector3 &normal, double &samplePdf) override;
         };
     }
 }
