+++ include/kaguya/math/RefractPdf.h
@@ -16,7 +16,7 @@ namespace kaguya {
 
             virtual double pdf(const Vector3 &inDir, const Vector3 &normal, const Vector3 &outDir) override;
 
-            virtual Vector3 random(const Vector3 &inDir, const Vector3 &normal, float &samplePdf) override;
+            virtual Vector3 random(const Vector3 &inDir, const Vector3 &normal, double &samplePdf) override;
 
         private:
             float _refractiveIndex;
